function that copies a string

+++ 0x06-pointers_arrays_strings/2-strncpy.c
@@ -7,9 +7,9 @@
  * Return: pointeur result
  */
 
-char *strncpy(char *dest, const char *src, size_t n)
+char *strncpy(char *dest, const char *src, int n)
 {
-	size_t i;
+	int i;
 
 	for (i = 0; i < n && src[i] != '\0'; i++)
 		dest[i] = src[i];
